Duplicate entry (#57)
* Add route to duplicate time entry

* Allow user to duplicate an entry

* ui tweaks

* Fix bug that prevent user to remove entries without a project

We were expecting time_entry parameters to be sent during the
delete method when actually we only (need to) send the id

This issue only happen when the time_entry didin't have a project
associated.

* Fix bug when runnning individul mobile test

Running individual mobile test makes the test to fails during the
teardown as we didn't loaded the ApplicationSystemTestCase.

* Tweak the code